requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
statsmodels>=0.14.0
openai>=1.0.0
faiss-cpu>=1.7.4
